refactor(analysis): split source complexity mask
