+++ docs/architecture-consolidation-plan.md
@@ -80,7 +80,7 @@ fixtures:
 | Derived analysis report | `crates/tokmd-analysis/src/derived/mod.rs` and `crates/tokmd-analysis/src/derived/` | `mod.rs` 220; ratio owner 157; integrity owner 124; file metrics owner 119; language owner 88; distribution owner 80 | Keep `mod.rs` as the receipt assembly coordinator; distribution/histogram calculations, integrity hashing, density/ratio reports, top-file/offender reports, and language composition reports now live in coherent owner modules. Future derived work should start from concrete product or proof evidence rather than splitting for line count. |
 | Core facade and FFI | `crates/tokmd-core/src/lib.rs`, `crates/tokmd-core/src/tests.rs`, `crates/tokmd-core/src/mutation_tests.rs`, `crates/tokmd-core/src/receipts.rs`, `crates/tokmd-core/src/workflows/`, and `crates/tokmd-core/src/ffi/` | `lib.rs` 159; root facade test owner 137; mutation test owner 330; receipt helper owner 144; workflow coordinator 31; workflow support owner 99; language workflow owner 84; module workflow owner 94; export workflow owner 106; diff workflow owner 64; analysis workflow owner 169; analysis input owner 118; analysis request owner 178; cockpit workflow owner 90; FFI coordinator 98; FFI test owner 913; envelope owner 15; mode owner 149; input owner 187; parse owner 257; settings parser owner 150 | Language, module, export, diff, analysis, and cockpit workflow facades now live under `workflows/`; core receipt construction lives in `receipts.rs`; root facade tests and mutation-killing receipt/parser tests live in private test owners; shared workflow support helpers live in `workflows/support.rs`; analysis in-memory export preparation lives in `workflows/analyze/input.rs`; analysis request construction lives in `workflows/analyze/request.rs`; public `run_json` coordination, FFI tests, in-memory input decoding/path validation, strict JSON parsing, FFI settings construction, mode dispatch, and response-envelope conversion live under `ffi/`; remaining work is extracting any future FFI helpers without changing public APIs |
 | Python binding boundary | `crates/tokmd-python/src/lib.rs`, `crates/tokmd-python/src/runtime.rs`, `crates/tokmd-python/src/args.rs`, `crates/tokmd-python/src/envelope.rs`, and `crates/tokmd-python/src/tests.rs` | `lib.rs` 445; runtime owner 175; argument builder owner 59; envelope helper owner 19; binding proof owner 806 | Keep PyO3 wrapper functions and module registration in `lib.rs`; keep shared Python execution and JSON-module handling in `runtime.rs`; keep Python argument-dict construction in `args.rs`; keep envelope extraction and exception mapping in `envelope.rs`; keep Python binding contract tests in `tests.rs`. Preserve the `tokmd_python_binding` proof scope. |
-| Analysis complexity | `crates/tokmd-analysis/src/complexity/mod.rs` + `complexity/functions.rs` + `complexity/functions/` + `complexity/details.rs` + `complexity/details/` + `complexity/summary.rs` + `complexity/risk.rs` + `complexity/debt.rs` + `complexity/histogram.rs` + `complexity/language.rs` + `complexity/math.rs` + `complexity/tests/unit.rs` | 156 + 223 + Rust function owner 126 + 122 + Rust span owner 37 + Python span owner 79 + JS/TS span owner 57 + Go span owner 47 + C-family span owner 52 + 138 + 78 + 69 + 33 + 35 + 5 + 346 | Keep shared complexity logic in `tokmd-analysis`, with Rust function counting plus Rust, Python, JavaScript/TypeScript, Go, and C-family function-span detection in owner modules; split remaining language/source/summary helpers and local unit tests only when fresh evidence shows a clear next seam |
+| Analysis complexity | `crates/tokmd-analysis/src/complexity/mod.rs` + `complexity/functions.rs` + `complexity/functions/` + `complexity/details.rs` + `complexity/details/` + `complexity/summary.rs` + `complexity/risk.rs` + `complexity/debt.rs` + `complexity/histogram.rs` + `complexity/language.rs` + `complexity/math.rs` + `source_complexity.rs` + `source_complexity/mask.rs` + `complexity/tests/unit.rs` | 156 + 223 + Rust function owner 126 + 122 + Rust span owner 37 + Python span owner 79 + JS/TS span owner 57 + Go span owner 47 + C-family span owner 52 + 138 + 78 + 69 + 33 + 35 + review-source coordinator 264 + Rust mask owner 141 + 346 | Keep shared complexity logic in `tokmd-analysis`, with Rust function counting plus Rust, Python, JavaScript/TypeScript, Go, and C-family function-span detection in owner modules; keep review-oriented Rust source complexity in its coordinator and the comment/string/raw-string mask in `source_complexity/mask.rs`; split remaining language/source/summary helpers and local unit tests only when fresh evidence shows a clear next seam |
 | Analysis Halstead | `crates/tokmd-analysis/src/halstead/mod.rs`, `halstead/operators.rs`, `halstead/operators/sets.rs`, and `halstead/tokenizer.rs` | report coordinator 205; operator dispatch owner 34; operator set owner 310; tokenizer owner 116 | Keep aggregate report construction in `mod.rs`, language/operator lookup in `operators.rs`, static per-language operator sets in `operators/sets.rs`, and per-file token counting in `tokenizer.rs`. Preserve the internal `crate::halstead::*` test surface and keep changes under the `analysis_halstead` proof scope. |
 | Analysis maintainability | `crates/tokmd-analysis/src/maintainability/mod.rs`, `maintainability/index.rs`, and `maintainability/tests/` | Halstead attachment coordinator 129; index formula owner 65; behavior/proof tests in dedicated test modules | Keep maintainability index formula, grading, rounding, and formula tests in `index.rs`; keep Halstead metric attachment in `mod.rs`; preserve the internal `crate::maintainability::*` test surface and keep changes under the `analysis_maintainability` proof scope. |
 | Analysis near-duplicate detection | `crates/tokmd-analysis/src/near_dup/mod.rs`, `near_dup/clusters.rs`, `near_dup/fingerprint.rs`, and `near_dup/pairs.rs` | report coordinator 274; cluster owner 305; fingerprint owner 95; pair scoring owner 186 | Keep report assembly, eligibility filtering, fingerprint orchestration, and scope partitioning in `mod.rs`; keep cluster construction in `clusters.rs`; keep Winnowing tokenization, k-gram hashing, file reads, and fingerprint selection in `fingerprint.rs`; keep inverted indexing, shared-fingerprint counts, Jaccard scoring, and deterministic pair ordering in `pairs.rs`. Preserve the `analysis_near_dup` proof scope. |
